+++ src/enums.js
@@ -1,4 +1,29 @@
 /**
+    ### enums(map, [name])
+
+    Defines an enum of strings.
+
+    - `map` hash enum -> value
+    - `name` optional string useful for debugging
+
+    Example
+
+    ```javascript
+    var Direction = enums({
+        North: 0, 
+        East: 1,
+        South: 2, 
+        West: 3
+    });
+    ```
+
+    #### is(x)
+
+    Returns `true` if `x` belongs to the enum.
+
+    ```javascript
+    Direction.is('North'); // => true
+    ```
     #### enums.of(keys, [name])
 
     Returns an enums of an array of keys, useful when you don't mind to define
@@ -23,16 +48,16 @@ function enums(map, name) {
   name = name || 'enums()';
 
   function Enums(x) {
-    assert(Enums.is(x), 'bad %s', name);
     assert(!(this instanceof Enums), 'cannot use new with %s', name);
+    assert(Enums.is(x), 'bad %s', name);
+    // all enums types are idempotent
     return x;
   }
 
   Enums.meta = {
     kind: 'enums',
     map: map,
-    name: name,
-    ctor: false
+    name: name
   };
 
   Enums.is = function (x) {
@@ -44,10 +69,10 @@ function enums(map, name) {
 
 enums.of = function (keys, name) {
   keys = Str.is(keys) ? keys.split(' ') : keys;
-  var values = {};
+  var value = {};
   keys.forEach(function (k, i) {
-    values[k] = i;
+    value[k] = i;
   });
-  return enums(values, name);
+  return enums(value, name);
 };