+++ src/maybe.js
@@ -11,52 +11,25 @@
     Radio.is(null);    // => true
     Radio.is(1);       // => false
     ```    
-
-    ### enums(map, [name])
-
-    Defines an enum of strings.
-
-    - `map` hash enum -> value
-    - `name` optional string useful for debugging
-
-    Example
-
-    ```javascript
-    var Direction = enums({
-        North: 0, 
-        East: 1,
-        South: 2, 
-        West: 3
-    });
-    ```
-
-    #### is(x)
-
-    Returns `true` if `x` belongs to the enum.
-
-    ```javascript
-    Direction.is('North'); // => true
-    ```
 **/
 
-function maybe(Type, name) {
+function maybe(T, name) {
 
-  name = name || format('maybe(%s)', getName(Type));
+  name = name || format('maybe(%s)', getName(T));
 
-  function Maybe(values, mut) {
+  function Maybe(value, mut) {
     assert(!(this instanceof Maybe), 'cannot use new with %s', name);
-    return Nil.is(values) ? null : coerce(Type, values, mut);
+    return Nil.is(value) ? null : T(value, mut);
   }
 
   Maybe.meta = {
     kind: 'maybe',
-    type: Type,
-    name: name,
-    ctor: false // cannot use new with null
+    type: T,
+    name: name
   };
 
   Maybe.is = function (x) {
-    return Nil.is(x) || Type.is(x);
+    return Nil.is(x) || T.is(x);
   };
 
   return Maybe;